Rename gem + cleanup
* Renaming the gem from toddlerbot to slackify.
* Cleanup of `lib/slackify` folder.
* Added a new configuration: unhandled_handler. You can specify a subclass of `Slackify::Handlers::Base` in the config. This class will be called with `#unhandled` when a message has no regex match.
