+++ slackify.gemspec
@@ -1,20 +1,20 @@
 # frozen_string_literal: true
 
 Gem::Specification.new do |s|
-  s.name        = 'toddlerbot'
-  s.version     = '0.1.1'
+  s.name        = 'slackify'
+  s.version     = '0.1.2'
   s.date        = '2019-03-17'
-  s.summary     = 'Rails slackbot framework'
-  s.description = 'Slackbot framework for Rails using the Events API'
+  s.summary     = 'Slackbot framework for Rails using the Events API'
+  s.description = 'Slackbot framework for Rails using the Events API. Supports events, interactive messages and slash commands.'
   s.authors     = ['Justin Leger', 'Michel Chatmajian']
   s.email       = 'hey@justinleger.ca'
   s.files = Dir["{app,config,lib}/**/*", "Rakefile", "README.md", "CHANGELOG.md"]
-  s.homepage    = 'https://github.com/jusleg/toddlerbot'
+  s.homepage    = 'https://github.com/jusleg/slackify'
   s.license     = 'MIT'
 
   s.metadata = {
-    "source_code_uri" => "https://github.com/jusleg/toddlerbot",
-    "changelog_uri" => "https://github.com/jusleg/toddlerbot/blob/master/CHANGELOG.md"
+    "source_code_uri" => "https://github.com/jusleg/slackify",
+    "changelog_uri" => "https://github.com/jusleg/slackify/blob/master/CHANGELOG.md"
   }
 
   s.add_dependency 'rails'
